+++ performance-index/src/components/ActionBar.js
