+++ performance-index/src/components/Logo.js
