+++ performance-index/src/components/icons/UpArrowCircle.js
